requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile pyproject.toml -o requirements.txt
aiohappyeyeballs==2.4.6
    # via aiohttp
aiohttp==3.11.12
    # via instructor
aiosignal==1.3.2
    # via aiohttp
annotated-types==0.7.0
    # via pydantic
anyio==4.8.0
    # via
    #   httpx
    #   openai
attrs==25.1.0
    # via aiohttp
certifi==2025.1.31
    # via
    #   httpcore
    #   httpx
    #   requests
cfgv==3.4.0
    # via pre-commit
charset-normalizer==3.4.1
    # via requests
click==8.1.8
    # via typer
distlib==0.3.9
    # via virtualenv
distro==1.9.0
    # via openai
docstring-parser==0.16
    # via instructor
filelock==3.17.0
    # via virtualenv
frozenlist==1.5.0
    # via
    #   aiohttp
    #   aiosignal
h11==0.14.0
    # via httpcore
httpcore==1.0.7
    # via httpx
httpx==0.28.1
    # via
    #   left-right-brain-compat-bot (pyproject.toml)
    #   openai
identify==2.6.6
    # via pre-commit
idna==3.10
    # via
    #   anyio
    #   httpx
    #   requests
    #   yarl
iniconfig==2.0.0
    # via pytest
instructor==1.7.2
    # via swarm
jinja2==3.1.5
    # via instructor
jiter==0.8.2
    # via
    #   instructor
    #   openai
markdown-it-py==3.0.0
    # via rich
markupsafe==3.0.2
    # via jinja2
mdurl==0.1.2
    # via markdown-it-py
multidict==6.1.0
    # via
    #   aiohttp
    #   yarl
nodeenv==1.9.1
    # via pre-commit
numpy==2.2.2
    # via swarm
openai==1.61.1
    # via
    #   left-right-brain-compat-bot (pyproject.toml)
    #   instructor
    #   swarm
packaging==24.2
    # via pytest
platformdirs==4.3.6
    # via virtualenv
pluggy==1.5.0
    # via pytest
pre-commit==4.1.0
    # via swarm
propcache==0.2.1
    # via
    #   aiohttp
    #   yarl
pydantic==2.10.6
    # via
    #   instructor
    #   openai
pydantic-core==2.27.2
    # via
    #   instructor
    #   pydantic
pygments==2.19.1
    # via rich
pytest==8.3.4
    # via
    #   left-right-brain-compat-bot (pyproject.toml)
    #   swarm
python-dotenv==1.0.1
    # via left-right-brain-compat-bot (pyproject.toml)
pyyaml==6.0.2
    # via pre-commit
requests==2.32.3
    # via
    #   instructor
    #   swarm
rich==13.9.4
    # via
    #   instructor
    #   typer
shellingham==1.5.4
    # via typer
sniffio==1.3.1
    # via
    #   anyio
    #   openai
swarm @ git+https://github.com/openai/swarm.git@9db581cecaacea0d46a933d6453c312b034dbf47
    # via left-right-brain-compat-bot (pyproject.toml)
tenacity==9.0.0
    # via instructor
tqdm==4.67.1
    # via
    #   openai
    #   swarm
typer==0.15.1
    # via instructor
typing-extensions==4.12.2
    # via
    #   anyio
    #   openai
    #   pydantic
    #   pydantic-core
    #   typer
urllib3==2.3.0
    # via requests
virtualenv==20.29.1
    # via pre-commit
yarl==1.18.3
    # via aiohttp
